Refactoring tests so that they actually work
