+++ src/main/webapp/app/exercises/shared/submission-policy/submission-policy-update.module.ts
@@ -1,9 +1,10 @@
 import { NgModule } from '@angular/core';
 import { ArtemisSharedModule } from 'app/shared/shared.module';
 import { SubmissionPolicyUpdateComponent } from 'app/exercises/shared/submission-policy/submission-policy-update.component';
+import { ArtemisSharedComponentModule } from 'app/shared/components/shared-component.module';
 
 @NgModule({
-    imports: [ArtemisSharedModule],
+    imports: [ArtemisSharedModule, ArtemisSharedComponentModule],
     declarations: [SubmissionPolicyUpdateComponent],
     exports: [SubmissionPolicyUpdateComponent],
 })
