+++ src/test/javascript/spec/component/programming-exercise/programming-exercise-update.component.spec.ts
@@ -22,7 +22,6 @@ import {
     ProgrammingLanguageFeature,
     ProgrammingLanguageFeatureService,
 } from 'app/exercises/programming/shared/service/programming-language-feature/programming-language-feature.service';
-import { LockRepositoryPolicy, SubmissionPenaltyPolicy, SubmissionPolicyType } from 'app/entities/submission-policy.model';
 import { MockComponent, MockDirective, MockModule, MockPipe } from 'ng-mocks';
 import { NgxDatatableModule } from '@swimlane/ngx-datatable';
 import { AlertComponent } from 'app/shared/alert/alert.component';
@@ -41,7 +40,14 @@ import { TeamConfigFormGroupComponent } from 'app/exercises/shared/team-config-f
 import { DifficultyPickerComponent } from 'app/exercises/shared/difficulty-picker/difficulty-picker.component';
 import { RemoveAuxiliaryRepositoryButtonComponent } from 'app/exercises/programming/manage/update/remove-auxiliary-repository-button.component';
 import { ArtemisTranslatePipe } from 'app/shared/pipes/artemis-translate.pipe';
-import { NgForm, NumberValueAccessor, NgModel, CheckboxControlValueAccessor, DefaultValueAccessor, SelectControlValueAccessor, ReactiveFormsModule } from '@angular/forms';
+import {
+    NgForm,
+    NumberValueAccessor,
+    NgModel,
+    CheckboxControlValueAccessor,
+    DefaultValueAccessor,
+    SelectControlValueAccessor,
+} from '@angular/forms';
 import { IncludedInOverallScorePickerComponent } from 'app/exercises/shared/included-in-overall-score-picker/included-in-overall-score-picker.component';
 import { CategorySelectorComponent } from 'app/shared/category-selector/category-selector.component';
 import { AddAuxiliaryRepositoryButtonComponent } from 'app/exercises/programming/manage/update/add-auxiliary-repository-button.component';
@@ -63,9 +69,6 @@ import { ProgrammingExerciseInstructionAnalysisComponent } from 'app/exercises/p
 describe('ProgrammingExercise Management Update Component', () => {
     const courseId = 1;
     const course = { id: courseId } as Course;
-    const lockRepositoryPolicy = { type: SubmissionPolicyType.LOCK_REPOSITORY, submissionLimit: 5 } as LockRepositoryPolicy;
-    const submissionPenaltyPolicy = { type: SubmissionPolicyType.SUBMISSION_PENALTY, submissionLimit: 5, exceedingPenalty: 50.4 } as SubmissionPenaltyPolicy;
-    const brokenPenaltyPolicy = { type: SubmissionPolicyType.SUBMISSION_PENALTY } as SubmissionPenaltyPolicy;
 
     let comp: ProgrammingExerciseUpdateComponent;
     let fixture: ComponentFixture<ProgrammingExerciseUpdateComponent>;
@@ -77,9 +80,8 @@ describe('ProgrammingExercise Management Update Component', () => {
 
     beforeEach(() => {
         TestBed.configureTestingModule({
-            imports: [ArtemisTestModule, MockModule(NgxDatatableModule), MockModule(NgbModule), MockModule(OwlDateTimeModule), MockModule(ReactiveFormsModule)],
+            imports: [ArtemisTestModule, MockModule(NgxDatatableModule), MockModule(NgbModule), MockModule(OwlDateTimeModule)],
             declarations: [
-                SubmissionPolicyUpdateComponent,
                 ProgrammingExerciseUpdateComponent,
                 MockComponent(AlertComponent),
                 MockComponent(AlertErrorComponent),
@@ -248,127 +250,6 @@ describe('ProgrammingExercise Management Update Component', () => {
             expect(comp.isSaving).toEqual(false);
             expect(comp.programmingExercise).toEqual(expectedProgrammingExercise);
             expect(comp.isExamMode).toBeFalsy();
-        }));
-    });
-
-    describe('submission policy change', () => {
-        let expectedProgrammingExercise: ProgrammingExercise;
-
-        beforeEach(() => {
-            expectedProgrammingExercise = new ProgrammingExercise(undefined, undefined);
-            expectedProgrammingExercise.course = course;
-            const route = TestBed.inject(ActivatedRoute);
-            route.params = of({ courseId });
-            route.url = of([{ path: 'new' } as UrlSegment]);
-            route.data = of({ programmingExercise: expectedProgrammingExercise });
-            jest.spyOn(programmingExerciseFeatureService, 'getProgrammingLanguageFeature').mockReturnValue(getProgrammingLanguageFeature(ProgrammingLanguage.JAVA));
-        });
-
-        it('Should set policy object on exercise', fakeAsync(() => {
-            comp.ngOnInit();
-            fixture.detectChanges();
-            tick();
-
-            expect(expectedProgrammingExercise.submissionPolicy).toBeUndefined();
-            const submissionPolicyTypeField = fixture.nativeElement.querySelector('#field_submissionPolicy');
-
-            for (const type of [SubmissionPolicyType.LOCK_REPOSITORY, SubmissionPolicyType.SUBMISSION_PENALTY]) {
-                submissionPolicyTypeField.value = type;
-                submissionPolicyTypeField.dispatchEvent(new Event('change'));
-                fixture.detectChanges();
-                tick();
-
-                expect(expectedProgrammingExercise.submissionPolicy?.type).toBe(type);
-                expect(expectedProgrammingExercise.submissionPolicy?.id).toBeUndefined();
-            }
-
-            submissionPolicyTypeField.value = SubmissionPolicyType.NONE;
-            submissionPolicyTypeField.dispatchEvent(new Event('change'));
-            fixture.detectChanges();
-            tick();
-
-            expect(expectedProgrammingExercise.submissionPolicy?.type).toBe(SubmissionPolicyType.NONE);
-        }));
-
-        it('Should set submission limit correctly for all policy types', fakeAsync(() => {
-            comp.ngOnInit();
-            fixture.detectChanges();
-            tick();
-
-            const submissionPolicyTypeField = fixture.nativeElement.querySelector('#field_submissionPolicy');
-            for (const type of [SubmissionPolicyType.LOCK_REPOSITORY, SubmissionPolicyType.SUBMISSION_PENALTY]) {
-                submissionPolicyTypeField.value = type;
-                submissionPolicyTypeField.dispatchEvent(new Event('change'));
-                fixture.detectChanges();
-                tick();
-
-                const submissionLimitInputField = fixture.nativeElement.querySelector('#field_submissionLimit');
-                submissionLimitInputField.value = 5;
-                submissionLimitInputField.dispatchEvent(new Event('input'));
-                tick();
-
-                expect(expectedProgrammingExercise.submissionPolicy?.submissionLimit).toBe(5);
-            }
-        }));
-
-        it('Should set exceeding penalty correctly for submission penalty type', fakeAsync(() => {
-            comp.ngOnInit();
-            fixture.detectChanges();
-            tick();
-
-            const submissionPolicyTypeField = fixture.nativeElement.querySelector('#field_submissionPolicy');
-            submissionPolicyTypeField.value = SubmissionPolicyType.SUBMISSION_PENALTY;
-            submissionPolicyTypeField.dispatchEvent(new Event('change'));
-            fixture.detectChanges();
-            tick();
-
-            const submissionLimitExceededPenaltyInputField = fixture.nativeElement.querySelector('#field_submissionLimitExceededPenalty');
-            submissionLimitExceededPenaltyInputField.value = 73.73;
-            submissionLimitExceededPenaltyInputField.dispatchEvent(new Event('input'));
-            tick();
-
-            expect(expectedProgrammingExercise.submissionPolicy?.exceedingPenalty).toBe(73.73);
-        }));
-
-        it('Should display correct input fields when penalty policy is already set', fakeAsync(() => {
-            expectedProgrammingExercise.submissionPolicy = lockRepositoryPolicy;
-            comp.ngOnInit();
-            fixture.detectChanges();
-            tick();
-
-            const submissionPolicyTypeField = fixture.nativeElement.querySelector('#field_submissionPolicy');
-            const submissionLimitInputField = fixture.nativeElement.querySelector('#field_submissionLimit');
-
-            expect(submissionPolicyTypeField.value).toBe(SubmissionPolicyType.LOCK_REPOSITORY);
-            expect(submissionLimitInputField.value).toBe('5');
-        }));
-
-        it('Should display correct input fields when penalty policy is already set', fakeAsync(() => {
-            expectedProgrammingExercise.submissionPolicy = submissionPenaltyPolicy;
-            comp.ngOnInit();
-            fixture.detectChanges();
-            tick();
-
-            const submissionPolicyTypeField = fixture.nativeElement.querySelector('#field_submissionPolicy');
-            const submissionLimitInputField = fixture.nativeElement.querySelector('#field_submissionLimit');
-            const submissionLimitExceededPenaltyInputField = fixture.nativeElement.querySelector('#field_submissionLimitExceededPenalty');
-
-            expect(submissionPolicyTypeField.value).toBe(SubmissionPolicyType.SUBMISSION_PENALTY);
-            expect(submissionLimitInputField.value).toBe('5');
-            expect(submissionLimitExceededPenaltyInputField.value).toBe('50.4');
-        }));
-
-        it('Should display correct input fields when set policy is broken', fakeAsync(() => {
-            expectedProgrammingExercise.submissionPolicy = brokenPenaltyPolicy;
-            comp.ngOnInit();
-            fixture.detectChanges();
-            tick();
-
-            const submissionLimitInputField = fixture.nativeElement.querySelector('#field_submissionLimit');
-            const submissionLimitExceededPenaltyInputField = fixture.nativeElement.querySelector('#field_submissionLimitExceededPenalty');
-
-            expect(submissionLimitInputField.value).toBe('');
-            expect(submissionLimitExceededPenaltyInputField.value).toBe('');
         }));
     });
